Add basic model evaluation functionality
New file: 'modelevaluation.py' will load a saved model together with a specified number of testing images, then evaluate the models' performance on those testing images, creating a list of improperly classified images.  Finally, 'modelevaluation.py' will display the incorrectly classified images (next to a repr. sample of the bin the image was placed in) in a figure.

Several functions within 'classifier' were modified to accept additional arguments.

A new command line option is available for 'classifier.py': the user can specify the number of categories on which to train the model.