Add Experimentation w/ Image Enhancement
Add file `imageenhancement.py` that contains experimentation with various enhancement techniques.  File includes an ImageCollection class which acts as a container for various modifications/enhancements of a given image.  File also includes methods for displaying a grid of enhanced images.  Enhancements include: converting to grayscale and lab, equalizing each type of image using `cv2.equalizeHist`, and using CLAHE (contrast limited histogram equalization).

A method for instantiating many CLAHE objects (with varying hyperparameters of `gridSize` and `clipLimit`), then applying all such objects to a single image and depicting the images in a grid is also included.

The next step in this direction will be a systematic tuning of these parameters on a fixed subset of the entire gtsrb database.